Use uri-reference-kmp in commonMain. Remove platform-specific implementations.
